+++ backend/.gen/reelgo/public/model/trips.go
@@ -9,6 +9,6 @@ package model
 
 type Trips struct {
 	UID      int32 `sql:"primary_key"`
-	Tid      int32 `sql:"primary_key"`
+	Tid      int64 `sql:"primary_key"`
 	TripName string
 }
